Перенес render и createElement в один модуль. Планирую переписать в компонентах createTemplate без использования вложенных шаблонов.
