+++ src/components/trip-filter.js
@@ -1,4 +1,4 @@
-import createElement from "../utils/create-element.js";
+import {createElement} from "../utils/render.js";
 
 const createFilter = (name) => {
   return {type: name.toLowerCase(), name};
