+++ src/main.js
@@ -6,26 +6,13 @@ import TripComponent from "./components/trip.js";
 import backend from "./data/backend.js";
 import createTrip from "./data/trip.js";
 import constants from "./data/constants.js";
+import {render} from "./utils/render.js";
 
 const sitePageHeaderElement = document.querySelector(`.page-header`);
 const tripMainElement = sitePageHeaderElement.querySelector(`.trip-main`);
 const tripViewHeaderElement = tripMainElement.querySelector(`.trip-controls .trip-view-header`);
 const tripFilterHeaderElement = tripMainElement.querySelector(`.trip-controls .trip-filter-header`);
 const tripEventsElement = document.querySelector(`main .trip-events`);
-
-const render = (container, element, place) => {
-  switch (place) {
-    case constants.RENDER_POSITIONS.AFTERBEGIN:
-      container.prepend(element);
-      break;
-    case constants.RENDER_POSITIONS.BEFOREEND:
-      container.append(element);
-      break;
-    case constants.RENDER_POSITIONS.AFTEREND:
-      container.after(element);
-      break;
-  }
-};
 
 const renderAll = () => {
   const trip = createTrip(backend.getPoints());
