--- src/utils/create-element.js
@@ -1,8 +1,0 @@
-const createElement = (template) => {
-  const newElement = document.createElement(`div`);
-  newElement.innerHTML = template;
-
-  return newElement.firstChild;
-};
-
-export default createElement;
